Some refactoring and put commands into its own class
